requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Async HTTP
httpx==0.26.0
aiohttp==3.9.1

# Database
supabase==2.3.4
asyncpg==0.29.0

# Microsoft Graph API
msal==1.26.0
azure-identity==1.15.0

# OpenAI
openai==1.10.0

# HTML/PDF processing
beautifulsoup4==4.12.3
lxml==5.1.0
weasyprint==60.2
jinja2==3.1.3

# Date/time parsing
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
